Fix date mapping

+++ app/src/main/java/com/alextos/cashback/core/data/dto/PaymentDto.kt
@@ -6,6 +6,6 @@ import kotlinx.serialization.Serializable
 data class PaymentDto(
     val id: String,
     val amount: Int,
-    val date: Long,
+    val date: Double,
     val source: CardDto? = null
 )
